Move form builder to Playbook::Forms::Builder
