--- playbook/app/pb_kits/playbook/pb_form/form_builder/typeahead_field.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-module Playbook
-  module PbForm
-    class FormBuilder
-      module TypeaheadField
-        def typeahead(name, _options = {}, _html_options = {}, props: {})
-          props[:name] = name
-          @template.pb_rails("typeahead", props: props)
-        end
-      end
-    end
-  end
-end
