+++ playbook/lib/playbook.rb
@@ -6,8 +6,10 @@ require "webpacker"
 require "webpacker/react"
 require "view_component/engine"
 
-require "playbook/props"
 require "playbook/version"
+
+require "playbook/props"
+require "playbook/forms"
 require "playbook/pb_kit_helper"
 require "playbook/pb_doc_helper"
 require "playbook/kit_base"
