+++ playbook/lib/playbook/forms.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require_relative "forms/builder"
